create common paging handler
Extract the paging logic into its own handler - this makes it reusable for the soon-to-be-added batch retry route
